Update README.md
TASK 1:Create a program that generates a random number and asks the user to guess it. Provide feedback on whether the guess is too high or too low until the user guesses the correct number.

TASK 2:Develop a calculator program that performs basic arithmetic operations such as addition, subtraction, multiplication, and division. Allow the user to input two numbers and choose an operation to perform.

TASK 3:Build a simple console-based Tic-Tac-Toe game that
allows two players to play against each other